requirements.txt
Django>=3.2.3,<3.3
django-smart-selects==1.5.9
djangorestframework==3.13.1
requests==2.28.1
#telegram-send==0.25
django-admin-interface==0.18.3
django-celery-beat==2.5.0
celery==5.2.7
redis==4.5.1
flower==1.2.0
psycopg2==2.9.5
gunicorn==20.1.0
requests_toolbelt==0.10.1
paho-mqtt==1.6.1
configparser
django-dump-load-utf8==0.0.4
django-tgbot==2.0.0
gunicorn